requirements.txt
Flask
Flask-Session
psycopg2-binary
SQLAlchemy
flask_login
flask-wtf
